Add OAuth integration (beta cluster)
Load the `mediawiki/oauthclient-php` composer library to simplify
OAuth interactions.

Additionally, replace the footer at the bottom with a header at
the top, and add a link back to the main page to the header bar.

Error reporting is enabled for E_ALL, and a session is always started
for the user.

Change-Id: I4800195b8e4b9ed45f1171d432bdec8366f9c3e4